QCamera2: HAL1: Force Flash Off only when StillMore alone is enabled
Issue:
Flash is turned Off in SeeMore even though UI enabled it.
It's because, if SeeMore is enabled then StillMore is forcibly
enabled from HAL to get the same effect for LiveSnapshot.
But due to this Flash is turned Off assuming StillMore is
On.

Fix:
Set flash Off only when StillMore is enabled and SeeMore is Off.
By this Flash is controlled only when StillMore alone is enabled.

Change-Id: I4baf10377d5b91bd5a3042b2bb081ee447ad9b1a
CRs-Fixed: 1062032
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>